+++ lib/packageScripts/publishPackages.d.ts
@@ -1,3 +1,4 @@
-export declare function publishPackages({ cwd }: {
+export declare function publishPackages({ cwd, choice }: {
     cwd: string;
+    choice: boolean;
 }): Promise<void>;