+++ lib/packageScripts/selectPublishOptions.d.ts
@@ -1,4 +1,9 @@
 import { PackagePublishOption } from '../types';
-export declare function selectPublishOptions({ publishOptions }: {
+export declare function getVersions({ currentPackageJson, remotePackageJson }: PackagePublishOption): {
+    currentVersion: string;
+    remoteVersion: string | undefined;
+};
+export declare function selectPublishOptions({ publishOptions, choice }: {
     publishOptions: PackagePublishOption[];
+    choice: boolean;
 }): Promise<PackagePublishOption[]>;