+++ src/index.ts
@@ -1,4 +1,5 @@
 export { webappScripts } from './webappScripts';
 export { packageScripts } from './packageScripts';
+export { desktopappScripts } from './desktopappScripts';
 export { createWebstormWebpackConfig } from './createWebstormWebpackConfig';
 export { patchStorybookWebpackConfig } from './patchStorybookWebpackConfig';