+++ src/packageScripts/createPackageBuildOptions.test.ts
@@ -6,7 +6,7 @@ import { createPackageBuildOptions } from './createPackageBuildOptions';
 describe('createPackageBuildOptions', () => {
   test('기본 BuildOptions를 생성한다', async () => {
     const cwd: string = await createTmpFixture('packages');
-    const entry: string[] = await getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
+    const entry: string[] = getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
     
     await expect(createPackageBuildOptions({entry, cwd})).resolves.toEqual([
       {
@@ -32,7 +32,7 @@ describe('createPackageBuildOptions', () => {
   
   test('기본 BuildOptions를 생성한다 (@group)', async () => {
     const cwd: string = await createTmpFixture('packages-group');
-    const entry: string[] = await getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
+    const entry: string[] = getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
     
     await expect(createPackageBuildOptions({entry, cwd})).resolves.toEqual([
       {
@@ -58,14 +58,14 @@ describe('createPackageBuildOptions', () => {
   
   test('index 파일이 없으면 Error를 발생시킨다', async () => {
     const cwd: string = await createTmpFixture('packages-no-index');
-    const entry: string[] = await getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
+    const entry: string[] = getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
     
     await expect(createPackageBuildOptions({entry, cwd})).rejects.toThrow();
   });
   
   test('index 파일이 여러개면 Error를 발생시킨다', async () => {
     const cwd: string = await createTmpFixture('packages-multi-index');
-    const entry: string[] = await getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
+    const entry: string[] = getInternalPackageEntry({packageDir: path.join(cwd, 'src/_packages')});
     
     await expect(createPackageBuildOptions({entry, cwd})).rejects.toThrow();
   });