+++ src/runners/copyServerPackageJson.ts
@@ -4,5 +4,8 @@ import path from 'path';
 export async function copyServerPackageJson({file, copyTo}: {file: string, copyTo: string}) {
   const {dependencies} = await fs.readJson(file);
   await fs.mkdirp(path.dirname(copyTo));
-  await fs.writeJson(copyTo, {dependencies}, {encoding: 'utf8'});
+  const content: object = {
+    dependencies,
+  };
+  await fs.writeJson(copyTo, content, {encoding: 'utf8'});
 }