+++ src/webpackConfigs/getWebpackAlias.ts
@@ -1,5 +1,6 @@
 import fs from 'fs-extra';
 import path from 'path';
+import { getInternalPackageEntry } from '../internalPackage/getInternalPackageEntry';
 
 export function getWebpackAlias({cwd}: {cwd: string}): {[moduleName: string]: string} {
   const src: string = path.join(cwd, 'src');
@@ -13,10 +14,9 @@ export function getWebpackAlias({cwd}: {cwd: string}): {[moduleName: string]: st
     .forEach((dirpath: string) => alias[path.basename(dirpath)] = path.resolve(cwd, dirpath));
   
   if (fs.existsSync(packageDir) && fs.statSync(packageDir).isDirectory()) {
-    fs.readdirSync(path.join(src, '_packages'))
-      .map((dirname: string) => path.join(src, `_packages/${dirname}`))
-      .filter((dirpath: string) => fs.statSync(dirpath).isDirectory())
-      .forEach((dirpath: string) => alias[path.basename(dirpath)] = path.resolve(cwd, dirpath));
+    getInternalPackageEntry({packageDir: path.join(src, '_packages')})
+      .filter((packageName: string) => fs.statSync(path.join(src, `_packages/${packageName}`)).isDirectory())
+      .forEach((packageName: string) => alias[packageName] = path.join(src, `_packages/${packageName}`));
   }
   
   return alias;